+++ shared/modules/utility_module.go
@@ -7,22 +7,28 @@ import (
 	"google.golang.org/protobuf/types/known/anypb"
 )
 
-const UtilityModuleName = "utility"
+const (
+	UtilityModuleName = "utility"
+)
+
+// TECHDEBT: Replace []byte with semantic type (addresses, transactions, etc...)
 
 type UtilityModule interface {
 	Module
 
-	// General purpose handler of utility specific messages that are not externalized in shared directories
-	HandleMessage(*anypb.Any) error
-
-	// Creates a `utilityContext` with an underlying read-write `persistenceContext`; only 1 of which can exist at a time
+	// NewContext creates a `utilityContext` with an underlying read-write `persistenceContext` (only 1 of which can exist at a time)
 	NewContext(height int64) (UtilityContext, error)
 
-	// Basic Transaction validation.
-	// SIDE EFFECT: Transaction is added to the utility's module mempool if valid to be repeated in the future; not obvious from the functional name.
-	CheckTransaction(tx []byte) error
+	// HandleTransaction does basic `Transaction` validation & adds it to the utility's module mempool if valid
+	HandleTransaction(tx []byte) error
 
+	// GetMempool returns the utility module's mempool of transactions gossiped throughout the network
 	GetMempool() mempool.TXMempool
+
+	// HandleUtilityMessage is a general purpose handler of utility-specific messages used for utility-specific business logic.
+	// It is useful for handling messages from the utility module's of other nodes that do not directly affect the state.
+	// IMPROVE: Find opportunities to break this apart as the module matures.
+	HandleUtilityMessage(*anypb.Any) error
 }
 
 // TECHDEBT: `CreateAndApplyProposalBlock` and `ApplyBlock` should be be refactored into a
@@ -30,33 +36,32 @@ type UtilityModule interface {
 
 // The context within which the node can operate with the utility layer.
 type UtilityContext interface {
-	// Block operations
-
-	// This function is intended to be called by any type of node during state transitions.
-	// For example, both block proposers and replicas/verifiers will use it to create a
-	// context (before finalizing it) during consensus, and all verifiers will call it during state sync.
-	// TODO: Replace []byte with semantic type
+	// SetProposalBlock updates the utility context with the proposed state transition.
+	// It does not apply, validate or commit the changes.
+	// For example, it can be use during state sync to set a proposed state transition before validation.
+	// TODO: Investigate a way to potentially simplify the interface by removing this function.
 	SetProposalBlock(blockHash string, proposerAddr []byte, txs [][]byte) error
-	// Reaps the mempool for transactions to be proposed in a new block, and applies them to this
-	// context.
+
+	// CreateAndApplyProposalBlock reaps the mempool for txs to be proposed in a new block, and
+	// applies them to this context after validation.
 	// Only intended to be used by the block proposer.
-	// TODO: Replace []byte with semantic type
 	CreateAndApplyProposalBlock(proposer []byte, maxTxBytes int) (stateHash string, txs [][]byte, err error)
-	// Applies the proposed local state (i.e. the transactions in the current context).
+
+	// ApplyBlock applies the context's in-memory proposed state (i.e. the txs in this context).
 	// Only intended to be used by the block verifiers (i.e. replicas).
 	ApplyBlock() (stateHash string, err error)
 
-	// Context operations
-
-	// Releases the utility context and any underlying contexts it references
+	// Release releases this utility context and any underlying contexts it references
 	Release() error
-	// Commit the current utility context (along with its underlying persistence context) to disk
+
+	// Commit commits this utility context along with any underlying contexts (e.g. persistenceContext) it references
 	Commit(quorumCert []byte) error
-	// Returns the read-write persistence context initialized by this utility context
-	GetPersistenceContext() PersistenceRWContext
 }
 
-// TECHDEBT: Remove this interface from `shared/modules`
+// TECHDEBT: Remove this interface from `shared/modules` and use the `Actor` protobuf type instead
+// There will need to be some documentation or indicator that the Actor struct returned may not be
+// fully hydrated. Alternatively, we could eat the performance cost and just hydrate the entire struct
+// which may be simpler and clearer.
 type UnstakingActor interface {
 	GetAddress() []byte
 	GetOutputAddress() []byte