+++ utility/transaction_test.go
@@ -7,9 +7,10 @@ import (
 
 	"github.com/pokt-network/pocket/runtime/test_artifacts"
 	"github.com/pokt-network/pocket/shared/codec"
-	"github.com/pokt-network/pocket/shared/converters"
+	"github.com/pokt-network/pocket/shared/core/types"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/crypto"
+	"github.com/pokt-network/pocket/shared/utils"
 	typesUtil "github.com/pokt-network/pocket/utility/types"
 	"github.com/stretchr/testify/require"
 )
@@ -22,9 +23,9 @@ func TestUtilityContext_AnteHandleMessage(t *testing.T) {
 	ctx := newTestingUtilityContext(t, 0)
 
 	tx, startingBalance, _, signer := newTestingTransaction(t, ctx)
-	_, signerString, err := ctx.anteHandleMessage(tx)
+	msg, err := ctx.anteHandleMessage(tx)
 	require.NoError(t, err)
-	require.Equal(t, signer.Address().String(), signerString)
+	require.Equal(t, signer.Address().Bytes(), msg.GetSigner())
 	feeBig, err := ctx.getMessageSendFee()
 	require.NoError(t, err)
 
@@ -38,7 +39,7 @@ func TestUtilityContext_ApplyTransaction(t *testing.T) {
 	ctx := newTestingUtilityContext(t, 0)
 
 	tx, startingBalance, amount, signer := newTestingTransaction(t, ctx)
-	txResult, err := ctx.applyTx(0, tx)
+	txResult, err := ctx.hydrateTxResult(tx, 0)
 	require.NoError(t, err)
 	require.Equal(t, int32(0), txResult.GetResultCode())
 	require.Equal(t, "", txResult.GetError())
@@ -52,20 +53,18 @@ func TestUtilityContext_ApplyTransaction(t *testing.T) {
 	require.Equal(t, expectedAfterBalance, amount, "unexpected after balance")
 }
 
-func TestUtilityContext_CheckTransaction(t *testing.T) {
-	mockBusInTestModules(t)
-
+func TestUtilityContext_HandleTransaction(t *testing.T) {
 	ctx := newTestingUtilityContext(t, 0)
 	tx, _, _, _ := newTestingTransaction(t, ctx)
 
 	txBz, err := tx.Bytes()
 	require.NoError(t, err)
-	require.NoError(t, testUtilityMod.CheckTransaction(txBz))
+	require.NoError(t, testUtilityMod.HandleTransaction(txBz))
 
 	hash, err := tx.Hash()
 	require.NoError(t, err)
 	require.True(t, testUtilityMod.GetMempool().Contains(hash))
-	require.Equal(t, testUtilityMod.CheckTransaction(txBz).Error(), typesUtil.ErrDuplicateTransaction().Error())
+	require.Equal(t, testUtilityMod.HandleTransaction(txBz).Error(), typesUtil.ErrDuplicateTransaction().Error())
 }
 
 func TestUtilityContext_GetSignerCandidates(t *testing.T) {
@@ -73,7 +72,7 @@ func TestUtilityContext_GetSignerCandidates(t *testing.T) {
 	accs := getAllTestingAccounts(t, ctx)
 
 	sendAmount := big.NewInt(1000000)
-	sendAmountString := converters.BigIntToString(sendAmount)
+	sendAmountString := utils.BigIntToString(sendAmount)
 	addrBz, er := hex.DecodeString(accs[0].GetAddress())
 	require.NoError(t, er)
 	addrBz2, er := hex.DecodeString(accs[1].GetAddress())
@@ -87,15 +86,13 @@ func TestUtilityContext_GetSignerCandidates(t *testing.T) {
 }
 
 func TestUtilityContext_CreateAndApplyBlock(t *testing.T) {
-	mockBusInTestModules(t)
-
 	ctx := newTestingUtilityContext(t, 0)
 	tx, _, _, _ := newTestingTransaction(t, ctx)
 
-	proposer := getFirstActor(t, ctx, coreTypes.ActorType_ACTOR_TYPE_VAL)
+	proposer := getFirstActor(t, ctx, types.ActorType_ACTOR_TYPE_VAL)
 	txBz, err := tx.Bytes()
 	require.NoError(t, err)
-	require.NoError(t, testUtilityMod.CheckTransaction(txBz))
+	require.NoError(t, testUtilityMod.HandleTransaction(txBz))
 
 	appHash, txs, er := ctx.CreateAndApplyProposalBlock([]byte(proposer.GetAddress()), 10000)
 	require.NoError(t, er)
@@ -109,11 +106,11 @@ func TestUtilityContext_HandleMessage(t *testing.T) {
 	accs := getAllTestingAccounts(t, ctx)
 
 	sendAmount := big.NewInt(1000000)
-	sendAmountString := converters.BigIntToString(sendAmount)
-	senderBalanceBefore, err := converters.StringToBigInt(accs[0].GetAmount())
+	sendAmountString := utils.BigIntToString(sendAmount)
+	senderBalanceBefore, err := utils.StringToBigInt(accs[0].GetAmount())
 	require.NoError(t, err)
 
-	recipientBalanceBefore, err := converters.StringToBigInt(accs[1].GetAmount())
+	recipientBalanceBefore, err := utils.StringToBigInt(accs[1].GetAmount())
 	require.NoError(t, err)
 	addrBz, er := hex.DecodeString(accs[0].GetAddress())
 	require.NoError(t, er)
@@ -122,17 +119,17 @@ func TestUtilityContext_HandleMessage(t *testing.T) {
 	msg := NewTestingSendMessage(t, addrBz, addrBz2, sendAmountString)
 	require.NoError(t, ctx.handleMessageSend(&msg))
 	accs = getAllTestingAccounts(t, ctx)
-	senderBalanceAfter, err := converters.StringToBigInt(accs[0].GetAmount())
+	senderBalanceAfter, err := utils.StringToBigInt(accs[0].GetAmount())
 	require.NoError(t, err)
 
-	recipientBalanceAfter, err := converters.StringToBigInt(accs[1].GetAmount())
+	recipientBalanceAfter, err := utils.StringToBigInt(accs[1].GetAmount())
 	require.NoError(t, err)
 
 	require.Equal(t, sendAmount, big.NewInt(0).Sub(senderBalanceBefore, senderBalanceAfter), "unexpected sender balance")
 	require.Equal(t, sendAmount, big.NewInt(0).Sub(recipientBalanceAfter, recipientBalanceBefore), "unexpected recipient balance")
 }
 
-func newTestingTransaction(t *testing.T, ctx *utilityContext) (tx *typesUtil.Transaction, startingBalance, amountSent *big.Int, signer crypto.PrivateKey) {
+func newTestingTransaction(t *testing.T, ctx *utilityContext) (tx *coreTypes.Transaction, startingBalance, amountSent *big.Int, signer crypto.PrivateKey) {
 	amountSent = new(big.Int).Set(defaultSendAmount)
 	startingBalance = new(big.Int).Set(test_artifacts.DefaultAccountAmount)
 
@@ -145,11 +142,11 @@ func newTestingTransaction(t *testing.T, ctx *utilityContext) (tx *typesUtil.Tra
 	signerAddr := signer.Address()
 	require.NoError(t, ctx.setAccountAmount(signerAddr, startingBalance))
 
-	msg := NewTestingSendMessage(t, signerAddr, recipientAddr.Bytes(), converters.BigIntToString(amountSent))
+	msg := NewTestingSendMessage(t, signerAddr, recipientAddr.Bytes(), utils.BigIntToString(amountSent))
 	any, err := codec.GetCodec().ToAny(&msg)
 	require.NoError(t, err)
 
-	tx = &typesUtil.Transaction{
+	tx = &coreTypes.Transaction{
 		Msg:   any,
 		Nonce: testNonce,
 	}