+++ p2p/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.30] - 2023-02-28
+
+- Renamed package names and parameters to reflect changes in the rest of the codebase
+
 ## [0.0.0.29] - 2023-02-24
 
 - Update logger value references with pointers
@@ -60,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.0.19] - 2023-01-19
 
 - Rewrite `interface{}` to `any`
+
 ## [0.0.0.18] - 2023-01-11
 
 - Add a lock to the mempool to avoid parallel messages which has caused the node to crash in the past