+++ runtime/docs/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.24] - 2023-02-28
+
+- Rename `app_staking_adjustment` to `app_session_tokens_multiplier`
+- Remove `app_baseline_stake_rate`
+- Rename `keygenerator` to `keygen`
+
 ## [0.0.0.23] - 2023-02-21
 
 - Rename ServiceNode Actor Type Name to Servicer